Use GitHub API to list images
The previous solution required a full checkout of the repo in order to
determine which images had changed. There was a subtle bug in there, as
the checkout was not done with full depth. Only the most recent commit
was fetched. This led to situations where the image list generation
failed with "bad object" errors.

There's a more straightforward approach using the GitHub API and the
gh CLI. It allows for easy fetching of diffs and massaging them via
jq queries. The same can be done to get the file list of a repository,
eliminating the need for a checkout in both cases.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>